initial service rework
This is a major cleanup of the dinit-chimera suite, making it
more fine-grained, with cleaner dependencies, better naming,
more (and more fine-grained) targets and others. Once this effort
is done, the target names will be considered stable, and services
outside dinit-chimera will be able to fully rely on them. Other
service names are internal to dinit-chimera and may change as
needed.