init-local.target: run after pre-network but before network
This effectively runs init-local in parallel with network setup,
which is probably the most convenient place to have it run (stuff
like firewalls is up, and network is guaranteed once it's up, so
people can do stuff like simple static configuration in rc.local).

There will never be a perfect place to run it, so people with
specific needs can just create a special scripted service and have
it run whenever they like.