account for rtc localtime/utc in swclock
We share the guessing logic with hwclock, assuming UTC by default
unless set to localtime via /etc/hwclock or /etc/adjtime.

If the RTC is in localtime, convert it to UTC epoch first, as
the gettimeofday as well as settimeofday high level functions deal
strictly in UTC.

Also, if built with 32-bit time_t, give up if we encounter values
not representable in the time_t.