cgroups: set up controllers ahead of time when mounting
This is needed for container runtimes as enabling controllers for
the root cgroup may not be possible on-demand (particularly, the
cpu controller will not enable if there are any realtime-priority
processes around, and instead it will return EINVAL and that may
make container runtimes fail).

By enabling them all ahead of time, we guarantee that this will
succeed as only pid1 runs at that time.