mnt: don't pass implicit flags to mount helpers
When we unparse, we want to unparse from only explicitly passed
flags (with the exception of something like remount) in order to
avoid bringing in e.g. MS_SILENT which some mount helpers will
not recognize (of course, if that is parsed from the original
flags string, we should still pass it).